Moved MockEmailConfirmationStrategy to 'confirmation' package
